Move calculateChartData to its own file
Pass calculateChartData inputs and setChartData as parameters so it can
be imported and called in Calculator's useEffect hook.

Declare a new function that calls calculateChartData so that an event
listener can be added and removed. This solves the issue I was having
with trying to remove an anonymous arrow function.